backend/requirements.txt
fastapi
uvicorn
opencv-python
ffmpeg-python
torch